+++ src/create/README.md
@@ -1,7 +1,7 @@
 # A visual text editor in 2000 lines of code
 
-Trying to implement a visual text editor within the frontend in less than 2000 lines of code.
+Trying to implement a visual text editor in less than 2000 lines of code.
 
-Current line count: 1029 (code: 961)
+Current line count: 1302
 
-A detailed code walkthrough will follow ...
+Entry point: `Create.tsx`
